requirements.txt
alembic==0.8.9
click==6.6
dominate==2.3.1
Flask==0.11.1
Flask-Bootstrap==3.3.7.0
Flask-Login==0.4.0
Flask-Migrate==2.0.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing-0.6.2
Flask-WTF==0.13.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
SQLAlchemy==1.1.4
visitor==0.1.3
Werkzeug==0.11.11
WTForms==2.1
python-coveralls (2.9.1)
